# Conflicts:
#	core/src/main/java/hzt/collections/primitives/DoubleArrayList.java
#	core/src/main/java/hzt/collections/primitives/DoubleCollection.java
#	core/src/main/java/hzt/collections/primitives/DoubleListX.java
#	core/src/main/java/hzt/collections/primitives/IntCollection.java
#	core/src/main/java/hzt/collections/primitives/IntListX.java
#	core/src/main/java/hzt/collections/primitives/LongCollection.java
#	core/src/main/java/hzt/collections/primitives/LongListX.java
#	core/src/main/java/hzt/sequences/SequenceHelper.java
#	core/src/test/java/hzt/arrays/primitives/DoubleSortTest.java
#	core/src/test/java/hzt/arrays/primitives/IntSortTest.java
#	core/src/test/java/hzt/arrays/primitives/LongSortTest.java
#	core/src/test/java/hzt/sequences/primitives/DoubleSequenceTest.java
#	core/src/test/java/hzt/sequences/primitives/IntSequenceTest.java

+++ core/src/main/java/hzt/collections/primitives/DoubleMutableListX.java
@@ -39,7 +39,7 @@ public interface DoubleMutableListX extends DoubleListX, DoubleMutableCollection
         return asSequence().boxed().toMutableList();
     }
 
-    long set(int index, long value);
+    double set(int index, double value);
 
     @Override
     default boolean remove(double d) {
